test/crypto: improve readability of ZUC256 tests
There were some ZUC256 tests for hash generate and hash verify mixed in
with the ZUC128 tests, with no indication in the test name that they
were for ZUC256.

These are now grouped with other ZUC256 tests, and all ZUC256 hash test
names are modified to be more uniform, and indicate the digest length
as this can vary with ZUC256.

Signed-off-by: Ciara Power <ciara.power@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>