lib/bpf: Rename 'bpf_validate' to avoid potential conflict with libpcap
The library libpcap has their function 'bpf_validate' either so
there would be a multiple definition issue when linking with
librte_bpf.a and libpcap.a staticly.

You can reproduce this issue by 'meson build -Dprefer_static=true
-Denable_apps=test-pmd -Denable_drivers=net/af_xdp,net/af_packet'.
Notice you need to have a static version of libpcap to reproduce this.

In 2019 there was a patch reported the same issue but not applied:
https://inbox.dpdk.org/stable/2601191342CEEE43887BDE71AB9772580148A95BE2@irsmsx105.ger.corp.intel.com/T

Since 'bpf_validate' is an internal function, I think adding an 'rte'
prefix is not a good idea and rename it to 'bpf_do_validate' instead.

Signed-off-by: Martzki <mars14850@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>