drivers: remove implementation of Rx metadata negotiation
Presently, Rx metadata is sent to PMD by default, leading
to a performance drop as processing for the same in Rx path
takes extra cycles.

Hence, removing driver implementation of Rx metadata negotiation
and falling back to old implementation where mark actions are
tracked as part of the flow rule.

Signed-off-by: Hanumanth Pothula <hpothula@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>