eal: add option to put timestamp on console output
When debugging driver or startup issues, it is useful to have
a timestamp on each message printed. The messages in syslog
already have a timestamp, but often syslog is not available
during testing. The timestamp format was chosen to be the
same as the default Linux dmesg timestamp.

The messages are written using writev() which avoids unnecessary
buffering and ensures that if multiple threads are logging at
the same time that timestamp and message buffer don't get
interleaved.

Example:
[       0.000040] EAL: Probing VFIO support...

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>