crypto/qat: added cipher-crc cap check
A configuration item called qat_sym_cipher_crc_enable has been added.
When set, an LA bulk req message with combined cipher-crc will be sent
on startup to the QAT device. The response is checked to see if the
data returned matches the cipher text. If a match is determined the
cipher-crc capability bit is set to indicate support.

If cipher-crc offload is supported, the LA Bulk request will be
formatted correctly before being enqueued to the device.

Signed-off-by: Kevin O'Sullivan <kevin.osullivan@intel.com>
Signed-off-by: David Coyle <david.coyle@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>