vhost: fix constants to follow new naming convention
DPDK apps (e.g. dpdk-skeleton) output this name
during negotiation. But, it is not consistent when
debugging using QEMU as a front-end, for example.
This is because QEMU already follows new naming convention [1].

Some type names and variable names, such as VhostUserSlaveRequest,
are still in old naming convention. But, in this patch we
only focus on constants.

[1] https://qemu-project.gitlab.io/qemu/interop/vhost-user.html

Signed-off-by: Nobuhiro MIKI <nmiki@yahoo-corp.jp>
Acked-by: Stephen Hemminger <stephen@networkplumber.org>
Reviewed-by: Chenbo Xia <chenbo.xia@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>