crypto/cnxk: update number of max capabilities
To ensure that the end marker can fit into the capabilities array, the
max number of capabilities has been incremented.
Additionally, throw an error if max length is reached. It's critical
because `rte_cryptodev_capabilities` in dev_info relies on the end marker.
Without it, all capabilities scans will read memory outside of the array.

Signed-off-by: Volodymyr Fialko <vfialko@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>