net/mlx5/hws: shared context uses defaults from local context
Fix default miss behavior for shared resources, the problem
could happen in two cases:
When the table created, the default miss should go to the alias ft
that will direct it back to local context.

When the rule is without specific hit address and we are in a shared
RTC from here it should be redirected back to the local context.

Fixes: ce946c7d3999 ("net/mlx5/hws: support ibv context shared with local one")
Reviewed-by: Alex Vesker <valex@nvidia.com>
Signed-off-by: Erez Shitrit <erezsh@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>