config/arm: support ldapr on neoverse n1
Armv8.3 introduced a LDAPR instruction that has weaker guarantees,
just sufficient to implement memory_order_acq_rel. LDAPR does not
wait for previous STLR to complete. And if the same location is
used in both, it does not force the write to be globally observed.
The new instruction can avoid some costs associated with LDAR.

GCC13.0 has supported LDAPR.
For Armv8.3 and higher, LDAPR is default.
For Armv8.2 such as neoverse n1, LDAPR is supported with enabling
rcpc feature.
For Armv8.1 and lower, LDAPR is not supported.

Signed-off-by: Joyce Kong <joyce.kong@arm.com>
Reviewed-by: Ruifeng Wang <ruifeng.wang@arm.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>