net/mlx5/hws: fix shared context to use local defaults
Fix default miss behavior for shared resources, the problem
could happen in two cases:
When the table created, the default miss should go to the alias ft
that will direct it back to local context.

When the rule is without specific hit address and we are in a shared
RTC from here it should be redirected back to the local context.

Fixes: ce946c7d3999 ("net/mlx5/hws: support ibv context shared with local one")
Cc: stable@dpdk.org

Signed-off-by: Erez Shitrit <erezsh@nvidia.com>
Reviewed-by: Alex Vesker <valex@nvidia.com>
Acked-by: Matan Azrad <matan@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>