net/mlx5/hws: fix saving of jumbo tag for deletion
In cases with more than one STE we didn't update the whole tag
of the match jumbo STE, Instead of 11B we kept only 8B like in
regular STE.
The ste_attr that keeps that indication was not updated at the
last iteration.
Now we indicate it is a jumbo frame before saving the tag for
future deletion of it.

Fixes: 405242c52dd5 ("net/mlx5/hws: add rule object")
Cc: stable@dpdk.org

Signed-off-by: Erez Shitrit <erezsh@nvidia.com>
Reviewed-by: Alex Vesker <valex@nvidia.com>
Acked-by: Matan Azrad <matan@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>